chore: forge-lint Batch 5.1 — 5 redundant cast removals + 3 reclassified suppressions
First focused sub-PR from the Batch 5 triage CSV. Targets the 8 sites
my heuristic classified as REDUNDANT-REMOVE; on closer per-site
analysis, 5 are genuine no-op outer casts (safe to remove), 3 are
double-cast patterns where both casts are necessary (suppress with
rationale instead).

### Genuine redundant outer casts removed (5 sites)

- test/DepthTieredLtv.t.sol  lines 148, 149, 211, 467
    `uint256(uint256(PRICE_2000))` → `uint256(PRICE_2000)`
    PRICE_2000 is `int256 constant`; the inner cast is the actual
    int256→uint256 conversion, the outer uint256→uint256 is a no-op.
    The inner cast remains lint-flagged (sign-conversion) — that's
    a separate REVIEW-SAFECAST bucket fix; this PR only clears the
    outer-cast warnings here.

- test/SecondaryQuorumTest.t.sol line 148
    `int256(int256(CHAINLINK_PRICE_8DEC))` → `int256(CHAINLINK_PRICE_8DEC)`
    CHAINLINK_PRICE_8DEC is `uint256 constant`; the inner int256(...)
    does the actual uint→int conversion, the outer is int256→int256.

### Reclassified to REVIEW-CONTEXT — suppressed with inline rationale (3 sites)

- src/facets/OracleFacet.sol:909  `uint256(int256(value))`
    `value` is `int224` returned by the API3 callback; widen int224 →
    int256 → uint256 is the correct pattern, guarded by the
    `if (value <= 0)` check immediately above. Both casts required.

- src/libraries/LibInteractionRewards.sol lines 374, 396
    `uint256(int256(open) + delta)` — `open` is uint256, `delta` is
    int256; signed arithmetic on the delta requires the int256 cast,
    and the result is non-negative by the invariant that lifetime
    deltas can never drive `open` below 0 (gated earlier in call
    sites). Both casts required.

Net: 8 lint warnings cleared (5 by removal + 3 by suppression).

The CSV at `docs/internal/batch5-unsafe-typecast-triage.csv` is
informational — these per-site corrections amend the heuristic
classification of 3 of the 8 entries; the CSV stays as-shipped in
PR #263 for the audit trail.

+++ contracts/test/SecondaryQuorumTest.t.sol
@@ -144,7 +144,7 @@ contract SecondaryQuorumTest is Test {
         _mockFeedFull(mockEthUsdFeed, int256(2000e8), 8);
         // Wire mockAsset → mockFeed via Feed Registry.
         _mockRegistryFeed(mockAsset, mockFeed);
-        _mockFeedFull(mockFeed, int256(int256(CHAINLINK_PRICE_8DEC)), 8);
+        _mockFeedFull(mockFeed, int256(CHAINLINK_PRICE_8DEC), 8);
 
         // Default state: all three secondaries DISABLED (zero address).
         // Tests that need active secondaries call set*() to enable.
